testing: extract common test utils from session.rs
They can be used in other test modules later.
